food-recommend-api/requirements.txt
fastapi
uvicorn[standard]
pandas
sentence-transformers
scikit-learn
torch
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
